better makefile

+++ Makefile
@@ -1,10 +1,5 @@
 VERSION=$(shell python3 -c "import quadpy; print(quadpy.__version__)")
 
-# Make sure we're on the master branch
-ifneq "$(shell git rev-parse --abbrev-ref HEAD)" "master"
-$(error Not on master branch)
-endif
-
 default:
 	@echo "\"make publish\"?"
 
@@ -15,12 +10,15 @@ README.rst: README.md
 
 # https://packaging.python.org/distributing/#id72
 upload: setup.py README.rst
+	# Make sure we're on the master branch
+	@if [ "$(shell git rev-parse --abbrev-ref HEAD)" != "master" ]; then exit 1; fi
 	rm -f dist/*
 	python3 setup.py bdist_wheel --universal
 	gpg --detach-sign -a dist/*
 	twine upload dist/*
 
 tag:
+	@if [ "$(shell git rev-parse --abbrev-ref HEAD)" != "master" ]; then exit 1; fi
 	@echo "Tagging v$(VERSION)..."
 	git tag v$(VERSION)
 	git push --tags
